requirements.txt
Flask==2.2.2
requests==2.28.2
Pillow==9.3.0
